ended???

+++ Dean-s_office/src/main/java/polsl/take/deansoffice/dtos/GradeDto.java
@@ -25,5 +25,5 @@ public class GradeDto extends RepresentationModel<GradeDto> {
 	@NotNull(message = "Final grade can not be empty!")
 	@Min(value = 1, message = "Final grade must be at least 2 (1 is for subject condition!)")
 	@Max(value = 5, message = "Final grade must be less than or equal to 5!")
-	private int finalGrade;
+	private float finalGrade;
 }
